Migrate existing architecture decision
We have one architectural decision already held in confluence. I am
moving it here. I've made some slight changes to the content to make it
match the format used here but tried to keep it as close to the original
as possible.

I will make architectural decision #2 the decision to
use GitHub for architectural decisions.