ADR for using GitHub for user directory (#8)
* ADR for using GitHub for user directory

We've talked quite a lot about using GitHub as our user directory. This
architectural decision record sets out that decision more clearly
explaining why we are doing that and how it impacts the work we are
doing on the new cloud platform.

* Make more specific about identify standards

We want to use apps or clients that work with GitHub over standards compliant OIDC or SAML wherever possible.