Add our support process document
We've been having an ongoing discussion on our team support processes.
As part of that we wanted to write up the processes that the team use so
that we have a clear record of them that we can all commit to.

This is the document. I am pushing this bit so that other people can
contribute to the various sections. My plan is to get a version together
that we can all sign up to.