Add ISSUE_TEMPLATE.md
There is still a lot of confusion about what constitutes a good
ticket/issue/story. I've taken a recent example that Kalbir wrote and
templated it for future reference.  I do not believe it will be
available when creating tickets directly in waffle.io, but you can
always raise the ticket on the repo, or copy the markdown out of another
ticket.