Add our support process document (#45)
* Add our support process document

We've been having an ongoing discussion on our team support processes.
As part of that we wanted to write up the processes that the team use so
that we have a clear record of them that we can all commit to.

This is the document. I am pushing this bit so that other people can
contribute to the various sections. My plan is to get a version together
that we can all sign up to.

* Add section on "How does the support sub team work"

I've collated togehter the various responses on how the support sub team
should work to make this schedule-lite for the team, identifying the key
activities that people need to do throughout the day.

* Add section on #ask-cloud-platform

There is quite a lot of information to add on how the
`#ask-cloud-platform` channel should work. This covers things like:

* raising tickets
* how to communicate in the channel
* some tips for making announcements

* Add all sections of the support doc

This is the summary of all of the conversations we had on support. As a
first draft for review. I expect that we will have a high number of
changes in review.

* Fixed headings and links

As @stevemarshall pointed out, headings were inconsistent, some used a
question style, others a statement. I've standardised them on
statements.

Also there were 2-3 links missing, I've put those in now.

* Fixed links, formatting and on call times

There were a bunch of small fixes, main ones are:

* Make slack channels links and use bacticks for all of them
* Sort out times of on call
* Link to anchors later in the doc

These were all in response to comments.