Add section for impact on user docs (#402)
We talked in retro about how we sometimes miss updating a doc as part of
the story we are completing. This prompt should help us consider the
imapct on the user docs when we are raising the story.

I've also changed the example checklist at the bottom to make it a
little more representative of what we do.