Only git push if runbooks were changed
If the current commit to the cloud-platform repo does not affect
the runbooks directory, then if the last thing this script does
is to try and 'git push', the command will fail, and the overall
exit status of the script will be an error.

This will show as a build failure on CircleCI, which is
inappropriate, since it doesn't indicate an error at all - just
that nothing in the runbooks was affected by the current commit.

This change does not try to `git push` if the previous `git commit`
failed. In that case, the last thing this script does is execute
`true`, so there should not be a build failure in CircleCI.