Alter index numbers of existing articles
It looks like a lot of articles use 200 as the
index number. This makes it hard to control the
ordering of items.

This change moves a couple of articles out of the
way, so that I can put two articles about tf
state locking next to each other in the index
panel.