Fix the "View Source" link on runbooks pages
Because the middleman site for the runbooks is in
the "runbooks" sub-directory of this repo, the 
"View Source" links at the bottom of each compiled
page don't work. 

e.g. the link points to:

https://github.com/ministryofjustice/cloud-platform/blob/master/source/upgrade-cluster.html.md.erb

...when it should point to:

https://github.com/ministryofjustice/cloud-platform/blob/master/runbooks/source/upgrade-cluster.html.md.erb

(i.e. "master" -> "master/runbooks")

This commit adds a script which inserts the
correct URL as the "source_url" frontmatter data
item for every page, before the HTML files are
compiled. Then the GDS tech-docs gem will use that
field as the target of the "View Source" link on
the compiled page.