Copy to the root folder of the publishing repo
The github pages site is published from the repo:
https://github.com/ministryofjustice/cloud-platform-runbooks

Previously, we were publishing from "master/docs"
In order not to use "master" we have changed the
default branch of that repo to "gh-pages" but this
means we can only publish from the root folder, 
not from docs. This is a limitation of github
pages.