Add "/runbooks" page to github_branch
The runbooks source files are in the runbooks *sub-directory* of the
cloud-platform repo, not the repo root (which seems to be the assumption
that the tech-docs gem is making).

This PR changes the "github_branch" from "main" to "main/runbooks". This
is a bit of a hack, but it seems to work, and fixes the "View Source"
links in a local development instance of the site.