Replace the real-looking AWS access key with XXX
We had an AWS access key in this runbook page which looked like a real
access key, even though it was not valid. This means we get spammed by
GitGuardian every week, alerting us about a leaked credential. This is
annoying, and means that we'd be less likely to pay attention to a
genuine credential-leak alert message.

I'm hoping this PR will fix that.