Rename and re-structure "Replacing a failed master node" (#2343)
* Rename and re-structure "Replacing a failed master node"

To avoid the document being confused with the process of terminating a
master node it was necessary to rename it. The introduction also
expanded upon and made more instructive.

In addition to this, I changed some of the variables so they're more
consistent, and replaced the long winded ssh commands with `kubectl
exec` commands. These commands are amended versions of this
[document](https://kops.sigs.k8s.io/operations/etcd_administration/#direct-data-access).

* Spelling and format corrections

* Change h1 title to reference the current page title