Remove blank line above metadata
The runbooks site has been failing to publish since #2607 was merged.

It seems that the blank line at the top of the file causes the middleman build step, which generates the HTML files, to fail.

This change should fix that. I'm going to go ahead and merge this now, because nobody else is around.