Add a command-line method to delete DynamoDB lock
I got sick of all the faffing around required to delete a terraform
state lock via the AWS console, so I spent a little while figuring out a
command-line route.

This script is partially untested, in that I successfully used a
previous version with the namespace hard-coded. So, it works in
principle, but the string interpolation part is only tested as far as I
could test it without actually deleting a real record (i.e. I invoked it
to delete the record it had already deleted, and that seemed to work as
expected).