Delete concourse worker pv cleanup runbook
We no longer had to cleanup the concouse worker PV. 
The latest concourse has failedGracePeriod for garbage collection of failed builds.
And maxDaysToRetainBuildLogs to retain only 30 days logs.

The issue with pv filling up is no longer occuring. If the issue re-occur, it will need further investigation instead of cleanup.