Increase review period to 6 months
After discussing in retro we agreed to increase the review period for documentation to 6 months.  Our infrastructure doesn't change that rapidly and when it does we update documents, this check is an additional one to make sure things don't get out of date.  By increasing the review time period we decrease the burden of reviewing docs.